obs-ffmpeg: Expose ffmpeg_data_init and ffmpeg_data_free methods
